Add new commands 'stream' and 'wifi' to configure camera settings. (#53)
New commands 'stream' and 'wifi' allow control of the Streaming and WiFi functions of
the attached camera. These settings can also be written to and read from files.

For Streaming settings, the official tools should be used to allocate the appropriate
crypto-tokens, the other values can be changed at will. Ustream documentation explains
how to obtain the OAuth2 tokens, the others are available from user's profile.